+++ src/component/Button.js
@@ -1,4 +1,4 @@
-import css from "../css/Button.module.css";
+import css from "./Button.module.css";
 
 export function LoginButton() {
   return <button className={css.loginBtn}>로그인</button>;
