+++ src/component/Search.module.css
@@ -1,7 +1,7 @@
 .root {
   display: flex;
   width: 90%;
-  height: 2.4rem;
+  height: 4.3rem;
   padding: 1.5rem 1.6rem;
   gap: 1rem;
   border-radius: 1rem;
