+++ src/containers/Nav/Nav.js
@@ -1,7 +1,6 @@
-import logoImg from "../assets/img/logo.svg";
-import style from "../css/Nav.module.css";
-import "../css/color.css";
-import NavProfile from "./NavProfile";
+import logoImg from "../../assets/img/logo.svg";
+import style from "./Nav.module.css";
+import NavProfile from "../../component/NavProfile";
 
 function Nav() {
   return (
